+++ src/ex_stage.sv
@@ -251,7 +251,9 @@ module ex_stage #(
 
     assign lsu_data  = lsu_valid_i ? fu_data_i  : '0;
 
-    load_store_unit lsu_i (
+    load_store_unit #(
+      .Cfg ( Cfg )
+    ) lsu_i (
         .clk_i,
         .rst_ni,
         .flush_i,